requirements.txt
torch==2.2.0
torchvision
torchaudio
ultralytics==8.2.14
streamlit
pillow
numpy
